Group chat app
I have created this group chat app in which when user enters username, it will get stored in localstorage and page will redirect to message from. When use types a message and submits it, the message will get stored into a file along with username.Also it will show above the message from on webpage.